creating types folder
