+++ frontend/src/contexts/PasswordContext.tsx
@@ -1,18 +1,6 @@
 import { createContext } from "react"
 
-import { PwdProps } from "../providers/PasswordProvider"
-
-export type PwdPreferences = {
-    name: string,
-    content: boolean | string,
-}
-
-type PwdContextProps = {
-    pwdData: PwdProps,
-    handlePwdUserPreferences: ({ name, content }: PwdPreferences) => void,
-    password: string
-    createPwd: () => void,
-}
+import { PwdContextProps } from "../types"
 
 export const initialValues = {
     pwdData: {
