+++ frontend/src/hooks/usePwd.ts
@@ -1,7 +1,7 @@
 "use client"
 
-import { useContext } from 'react'
+import { useContext } from "react"
 
-import { PasswordContext } from '../contexts/PasswordContext'
+import { PasswordContext } from "../contexts/PasswordContext";
 
 export const usePwd = () => useContext(PasswordContext)
